[ADDED] some nodes

+++ source/brdfs/GlossySpecular.cpp
@@ -19,9 +19,9 @@ GlossySpecular::GlossySpecular()
 RGBColor GlossySpecular::f(const ShadeRec& sr, const Vector3D& wo, const Vector3D& wi) const
 {
 	RGBColor	L;
-	float		ndotwi = sr.normal * wi;
-	Vector3D 	r(-wi + 2.0 * sr.normal * ndotwi); // mirror reflection direction
-	float		rdotwo = r * wo;
+	float		ndotwi = static_cast<float>(sr.normal * wi);
+	Vector3D 	r(-wi + 2.0f * sr.normal * ndotwi); // mirror reflection direction
+	float		rdotwo = static_cast<float>(r * wo);
 
 	if (rdotwo > 0.0)
 		L = m_ks * m_cs * pow(rdotwo, m_exp);
@@ -36,8 +36,8 @@ RGBColor GlossySpecular::rho(const ShadeRec& sr, const Vector3D& wo) const
 
 RGBColor GlossySpecular::sample_f(const ShadeRec& sr, const Vector3D& wo, Vector3D& wi, float& pdf) const
 {
-	float ndotwo = sr.normal * wo;
-	Vector3D r = -wo + 2.0 * sr.normal * ndotwo;	// direction of mirror reflection
+	float ndotwo = static_cast<float>(sr.normal * wo);
+	Vector3D r = -wo + 2.0f * sr.normal * ndotwo;	// direction of mirror reflection
 
 	Vector3D w = r;
 	Vector3D u = Vector3D(0.00424, 1, 0.00764) ^ w;
@@ -51,7 +51,7 @@ RGBColor GlossySpecular::sample_f(const ShadeRec& sr, const Vector3D& wo, Vector
 		wi = -sp.x * u - sp.y * v + sp.z * w;
 
 	float phong_lobe = pow((float)(r * wi), (float)m_exp);
-	pdf = phong_lobe * (sr.normal * wi);
+	pdf = static_cast<float>(phong_lobe * (sr.normal * wi));
 
 	return (m_ks * m_cs * phong_lobe);
 }
