+++ source/mapping/SphericalMap.cpp
@@ -13,13 +13,13 @@ void SphericalMap::GetTexelCoordinates(const Point3D& local_hit_point, int xres,
 	float theta = acos(local_hit_point.y);
 	float phi   = atan2(local_hit_point.x, local_hit_point.z);
 	if (phi < 0.0) {
-		phi += TWO_PI;
+		phi += static_cast<float>(TWO_PI);
 	}
 
 	// next, map theta and phi to (u, v) in [0, 1] X [0, 1]
 
-	float u = phi * INV_TWO_PI;
-	float v = 1.0 - theta * INV_PI;
+	float u = static_cast<float>(phi * INV_TWO_PI);
+	float v = static_cast<float>(1.0 - theta * INV_PI);
 
 	// finally, map u and v to the texel coordinates
